requirements.txt
# Story Engine Dependencies

# Database
psycopg2-binary==2.9.9

# Redis / Queue
redis==5.0.4

# Reddit
praw==7.7.1

# RSS / Web scraping
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.2

# AI Clients
anthropic==0.29.0
openai==1.35.0

# Language detection
langdetect==1.0.9

# YouTube upload
google-api-python-client==2.134.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Scheduler
APScheduler==3.10.4

# Utilities
python-dotenv==1.0.1